USB: convert khubd to a workqueue
Simply a cleanup as khubd open codes several facilities that are
provided by workqueue.  Note that the body of hub_events() is kept as a
loop to minimize thrash and help the readability of the patch.

This does reorder the actions in hub_quiesce().  Since the
->disconnected flag is removed we wait to flush the workqueue until
after the urb has been killed, and we know that kick_khubd() will not be
called again.  The kref for the usb_hub is also removed since we can
flush the workqueue and do not need to hold a reference while the hub
object while is being processed.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>