USB: make usb_ports proper parents of their child usb_devices
There are a number of places where pm_runtime manipulations are open
coded due to the fact that a usb_port is a device-model nephew of it's
"child" usb_device.  These can be cleaned up by making a usb_port the
device_model parent of its child.  With that relationship we rely on the
normal pm runtime rules that a parent device cannot be suspended until
all its children are suspended, and that attempts to resume children
will wake the parent first.

With this change we still want hubs to suspend when ports are inactive.
Previously ports would be forced to remain active when some power off
constraint was not met (like remote wakeup).  Now we let them suspend
when idle, but optionally keep the power enabled (return 0 from
usb_port_runtime_suspend).

In order to preserve the userspace abi, a link from the hub to the child
device is provided.

Note, the usb_device.parent remains the hub device.

Before:
# readlink -f /sys/devices/pci0000:00/0000:00:14.0/usb2/2-1
              /sys/devices/pci0000:00/0000:00:14.0/usb2/2-1
              +-------+
              |  hub  |
              +---^---+
                  |
            +-----+----+
            |          |
        +---+---+  +---+---+
        | intf  |  |device |
        +---^---+  +-------+
            |
        +---+---+
        | port  |
        +-------+

After:
# readlink -f /sys/devices/pci0000:00/0000:00:14.0/usb2/2-1
              /sys/devices/pci0000:00/0000:00:14.0/usb2/2-0:1.0/port1/2-1

              +-------+
              |  hub  |
              +---^---+
                  |
              +---+---+
              | intf  |
              +---^---+
                  |
              +---+---+
              | port  |
              +---^---+
                  |
              +---+---+
              |device |
              +-------+

Signed-off-by: Dan Williams <dan.j.williams@intel.com>