USB: kill unnecessary C_PORT_CONNECTION and broken C_PORT_ENABLE manipulations
C_PORT_CONNECTION changes are now handled and cleared while a port is
powered off.  It is also specified to be cleared by both usb2 [1] and
usb3 [2] hubs when the port is in the logical power-off state.

C_PORT_ENABLE is cleared by usb2 [3] hubs in the logical power-off
state, and the bit is reserved for usb3 [2] hubs.

In usb_port_runtime_resume() it does not make sense to attempt to
re-disable the port after a resume / debounce failure, we want the port
to be in enabled state regardless.

[1]: USB2.0 Specification Section 11.24.2.7.2.1
[2]: USB3.0 Specification Section 10.16.2.6.2
[3]: USB2.0 Specification Section 11.24.2.7.2.2

Signed-off-by: Dan Williams <dan.j.williams@intel.com>