usb: guarantee child device resume on port poweron
Make port power recovery behave similarly to the power session recovery
that occurs after a system / hub suspend event.  Arrange for a
usb_device to always complete a usb_port_resume() run prior to the next
khubd run.  This serves several purposes:

1/ The device may need a reset on power-session loss, without this
   change port power-on recovery exposes khubd to scenarios that
   usb_port_resume() is set to handle.  Also, testing showed that USB3
   devices that are not reset on power-session loss may eventually
   downgrade their connection to the USB2 pins.

2/ This mechanism rate limits port power toggling.  The minimum port
   power on/off period is now gated by the child device suspend/resume
   latency.  This mitigates devices downgrading their connection on
   perceived instability of the host connection.  This ratelimiting is
   really only relevant to port power control testing, but it is a nice
   side effect of closing the above race.

3/ Going forward if we find that power-session recovery requires
   warm-resets (http://marc.info/?t=138659232900003&r=1&w=2) that is
   something usb_port_resume() can drive and handle before khubd's next
   evaluation of the portstatus.

4/ If the device *was* disconnected the only time we'll know for sure is
   after a failed resume, so it's necessary for
   usb_port_runtime_resume() to expedite a usb_port_resume() to clean up
   the removed device.

1, 2, and 4 are not a problem in the system resume case because,
although the power-session is lost, khubd is frozen until after device
resume.  For the runtime pm case we can use runtime-pm-synchronization
to guarantee the same sequence of events.  When a ->resume_child request
is set in usb_port_runtime_resume() the port device is in the
RPM_RESUMING state.  khubd executes a pm_runtime_barrier() on the port
device to flush the port recovery, holds the port active while it
resumes the child, and completes child device resume before acting on
the current portstatus.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>