xhci: cancel in-flight resume requests when the port is powered off
Work around a remote wakeup vs port poweroff request race.  A wakeup request
sets the timer, but by the time it expires the port has been turned off, so we
hit:
                if ((raw_port_status & PORT_RESET) ||
                                !(raw_port_status & PORT_PE))
                        return 0xffffffff;

...in xhci_get_port_status

When userspace has set the port policy to allow poweroff the assumption is that
it no longer cares about remote wakeup requests for that port.  If a request
happens to collide with autosuspend of the port we need to make sure to cancel
an in-flight request otherwise we can end up in an infinite loop.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>