xhci: prepare for mid-segment link-trbs
Break the assumption that link-trbs are always located at the end of a
segment.  Introduce ->link to struct xhci_segment and use that every
place that looks up the link-trb for a segment.  This is meant to be
functionally equivalent to the existing driver and is just a search and
replace for hard coded "link at the end" assumption.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>