xhci: power-of-2 ring sizes
In preparation for dynamic ring expansion while walking scatterlists for
v1.0+ xhci hosts force the ring sizes to be a power-of-2.  This property
combined with the conversion of segments to a doubly linked list allows
for translating ring pointers to integers enabling simple math to
interrogate the state of the ring.  Beyond determining the amount of
free space in the ring this enables easy to calculate answers to
questions like:

* whether a trb is free or allocated?
* which segment precedes the current segment?
* how many trbs before the enqueue pointer wraps into the same segment
  as the dequeue pointer?

The open coded "num_trbs alignment" statement in xhci_ring_expansion()
is replaced with the standard ALIGN macro.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>